factor out base functionality of JSGF vs. GrXML to avoid GrXML deriving from JSGF.
fixes issue https://github.com/cmusphinx/sphinx4/issues/76
