+++ sphinx4-core/src/main/java/edu/cmu/sphinx/jsgf/JSGFGrammar.java
@@ -164,34 +164,12 @@ import edu.cmu.sphinx.util.props.S4String;
  * <li>All internal probabilities are maintained in LogMath log base.
  * </ol>
  */
-public class JSGFGrammar extends Grammar {
-
-    /** The property that defines the location of the JSGF grammar file. */
-    @S4String
-    public final static String PROP_BASE_GRAMMAR_URL = "grammarLocation";
-
-    /** The property that defines the location of the JSGF grammar file. */
-    @S4String(defaultValue = "default.gram")
-    public final static String PROP_GRAMMAR_NAME = "grammarName";
-
-    // ---------------------
-    // Configurable data
-    // ---------------------
-    private JSGFRuleGrammar ruleGrammar;
-    protected JSGFRuleGrammarManager manager;
-    protected RuleStack ruleStack;
-    private String grammarName;
-    protected URL baseURL;
-    private LogMath logMath;
-
-    protected boolean loadGrammar = true;
-    protected GrammarNode firstNode;
-    protected Logger logger;
+public class JSGFGrammar extends JSGFBaseGrammar {
 
     public JSGFGrammar(String location, String grammarName,
-            boolean showGrammar, boolean optimizeGrammar,
-            boolean addSilenceWords, boolean addFillerWords,
-            Dictionary dictionary) throws MalformedURLException,
+                           boolean showGrammar, boolean optimizeGrammar,
+                           boolean addSilenceWords, boolean addFillerWords,
+                           Dictionary dictionary) throws MalformedURLException,
             ClassNotFoundException {
         this(ConfigurationManagerUtils.resourceToURL(location),
                 grammarName, showGrammar, optimizeGrammar, addSilenceWords,
@@ -199,375 +177,14 @@ public class JSGFGrammar extends Grammar {
     }
 
     public JSGFGrammar(URL baseURL, String grammarName,
-            boolean showGrammar, boolean optimizeGrammar,
-            boolean addSilenceWords, boolean addFillerWords,
-            Dictionary dictionary) {
-        super(showGrammar, optimizeGrammar, addSilenceWords, addFillerWords,
-                dictionary);
-
-        logger = Logger.getLogger(getClass().getName());
-        logMath = LogMath.getLogMath();
-
-        this.baseURL = baseURL;
-        this.grammarName = grammarName;
-        loadGrammar = true;
+                           boolean showGrammar, boolean optimizeGrammar,
+                           boolean addSilenceWords, boolean addFillerWords,
+                           Dictionary dictionary) {
+        super(baseURL, grammarName, showGrammar, optimizeGrammar, addSilenceWords, addFillerWords, dictionary);
     }
 
     public JSGFGrammar() {
 
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see
-     * edu.cmu.sphinx.util.props.Configurable#newProperties(edu.cmu.sphinx.util
-     * .props.PropertySheet)
-     */
-    @Override
-    public void newProperties(PropertySheet ps) throws PropertyException {
-        super.newProperties(ps);
-        logger = ps.getLogger();
-        logMath = LogMath.getLogMath();
-
-        baseURL = ConfigurationManagerUtils.getResource(PROP_BASE_GRAMMAR_URL,
-                ps);
-        grammarName = ps.getString(PROP_GRAMMAR_NAME);
-
-        loadGrammar = true;
-    }
-
-    /**
-     * Returns the RuleGrammar of this JSGFGrammar.
-     * 
-     * @return the RuleGrammar
-     */
-    public JSGFRuleGrammar getRuleGrammar() {
-        return ruleGrammar;
-    }
-
-    /**
-     * Returns manager used to load grammars
-     * 
-     * @return manager with loaded grammars
-     */
-    public JSGFRuleGrammarManager getGrammarManager() {
-        if (manager == null)
-            manager = new JSGFRuleGrammarManager();
-        return manager;
-    }
-
-    /**
-     * Sets the URL context of the JSGF grammars.
-     * 
-     * @param url
-     *            the URL context of the grammars
-     */
-    public void setBaseURL(URL url) {
-        baseURL = url;
-    }
-
-    /** @return the name of this grammar. */
-    public String getGrammarName() {
-        return grammarName;
-    }
-
-    /**
-     * The JSGF grammar specified by grammarName will be loaded from the base
-     * URL (tossing out any previously loaded grammars)
-     * 
-     * @param grammarName
-     *            the name of the grammar
-     * @throws IOException
-     *             if an error occurs while loading or compiling the grammar
-     * @throws JSGFGrammarException
-     *              other grammar exception
-     * @throws JSGFGrammarParseException
-     *              failed to parse grammar file
-     */
-    public void loadJSGF(String grammarName) throws IOException,
-            JSGFGrammarParseException, JSGFGrammarException {
-        this.grammarName = grammarName;
-        loadGrammar = true;
-        commitChanges();
-    }
-
-    /**
-     * Creates the grammar.
-     * 
-     * @return the initial node of the Grammar
-     */
-    @Override
-    protected GrammarNode createGrammar() throws IOException {
-        try {
-            commitChanges();
-        } catch (JSGFGrammarException e) {
-            throw new IOException(e);
-        } catch (JSGFGrammarParseException e) {
-            throw new IOException(e);
-        }
-        return firstNode;
-    }
-
-    /**
-     * Returns the initial node for the grammar
-     * 
-     * @return the initial grammar node
-     */
-    @Override
-    public GrammarNode getInitialNode() {
-        return firstNode;
-    }
-
-    /**
-     * Parses the given Rule into a network of GrammarNodes.
-     * 
-     * @param rule
-     *            the Rule to parse
-     * @return a grammar graph
-     * @throws JSGFGrammarException exception during parse
-     */
-    protected GrammarGraph processRule(JSGFRule rule) throws JSGFGrammarException {
-        GrammarGraph result;
-
-        if (rule != null) {
-            logger.fine("parseRule: " + rule);
-        }
-
-        if (rule instanceof JSGFRuleAlternatives) {
-            result = processRuleAlternatives((JSGFRuleAlternatives) rule);
-        } else if (rule instanceof JSGFRuleCount) {
-            result = processRuleCount((JSGFRuleCount) rule);
-        } else if (rule instanceof JSGFRuleName) {
-            result = processRuleName((JSGFRuleName) rule);
-        } else if (rule instanceof JSGFRuleSequence) {
-            result = processRuleSequence((JSGFRuleSequence) rule);
-        } else if (rule instanceof JSGFRuleTag) {
-            result = processRuleTag((JSGFRuleTag) rule);
-        } else if (rule instanceof JSGFRuleToken) {
-            result = processRuleToken((JSGFRuleToken) rule);
-        } else {
-            throw new IllegalArgumentException("Unsupported Rule type: " + rule);
-        }
-        return result;
-    }
-
-    /**
-     * Parses the given RuleName into a network of GrammarNodes.
-     * 
-     * @param initialRuleName
-     *            the RuleName rule to parse
-     * @return a grammar graph
-     */
-    private GrammarGraph processRuleName(JSGFRuleName initialRuleName)
-            throws JSGFGrammarException {
-        logger.fine("parseRuleName: " + initialRuleName);
-        GrammarGraph result = ruleStack.contains(initialRuleName.getRuleName());
-
-        if (result != null) { // its a recursive call
-            return result;
-        } else {
-            result = new GrammarGraph();
-            ruleStack.push(initialRuleName.getRuleName(), result);
-        }
-        JSGFRuleName ruleName = ruleGrammar.resolve(initialRuleName);
-
-        if (ruleName == JSGFRuleName.NULL) {
-            result.getStartNode().add(result.getEndNode(), 0.0f);
-        } else if (ruleName == JSGFRuleName.VOID) {
-            // no connection for void
-        } else {
-            if (ruleName == null) {
-                throw new JSGFGrammarException("Can't resolve "
-                        + initialRuleName + " g "
-                        + initialRuleName.getFullGrammarName());
-            }
-            JSGFRuleGrammar rg = manager.retrieveGrammar(ruleName
-                    .getFullGrammarName());
-            if (rg == null) {
-                throw new JSGFGrammarException("Can't resolve grammar name "
-                        + ruleName.getFullGrammarName());
-            }
-
-            JSGFRule rule = rg.getRule(ruleName.getSimpleRuleName());
-            if (rule == null) {
-                throw new JSGFGrammarException("Can't resolve rule: "
-                        + ruleName.getRuleName());
-            }
-            GrammarGraph ruleResult = processRule(rule);
-            if (result != ruleResult) {
-                result.getStartNode().add(ruleResult.getStartNode(), 0.0f);
-                ruleResult.getEndNode().add(result.getEndNode(), 0.0f);
-            }
-        }
-        ruleStack.pop();
-        return result;
-    }
-
-    /**
-     * Parses the given RuleCount into a network of GrammarNodes.
-     * 
-     * @param ruleCount
-     *            the RuleCount object to parse
-     * @return a grammar graph
-     */
-    private GrammarGraph processRuleCount(JSGFRuleCount ruleCount)
-            throws JSGFGrammarException {
-        logger.fine("parseRuleCount: " + ruleCount);
-        GrammarGraph result = new GrammarGraph();
-        int count = ruleCount.getCount();
-        GrammarGraph newNodes = processRule(ruleCount.getRule());
-
-        result.getStartNode().add(newNodes.getStartNode(), 0.0f);
-        newNodes.getEndNode().add(result.getEndNode(), 0.0f);
-
-        // if this is optional, add a bypass arc
-
-        if (count == JSGFRuleCount.ZERO_OR_MORE
-                || count == JSGFRuleCount.OPTIONAL) {
-            result.getStartNode().add(result.getEndNode(), 0.0f);
-        }
-
-        // if this can possibly occur more than once, add a loopback
-
-        if (count == JSGFRuleCount.ONCE_OR_MORE
-                || count == JSGFRuleCount.ZERO_OR_MORE) {
-            newNodes.getEndNode().add(newNodes.getStartNode(), 0.0f);
-        }
-        return result;
-    }
-
-    /**
-     * Parses the given RuleAlternatives into a network of GrammarNodes.
-     * 
-     * @param ruleAlternatives
-     *            the RuleAlternatives to parse
-     * @return a grammar graph
-     */
-    private GrammarGraph processRuleAlternatives(
-            JSGFRuleAlternatives ruleAlternatives) throws JSGFGrammarException {
-        logger.fine("parseRuleAlternatives: " + ruleAlternatives);
-        GrammarGraph result = new GrammarGraph();
-
-        List<JSGFRule> rules = ruleAlternatives.getRules();
-        List<Float> weights = getNormalizedWeights(ruleAlternatives.getWeights());
-
-        // expand each alternative, and connect them in parallel
-        for (int i = 0; i < rules.size(); i++) {
-            JSGFRule rule = rules.get(i);
-            float weight = 0.0f;
-            if (weights != null) {
-                weight = weights.get(i);
-            }
-            logger.fine("Alternative: " + rule);
-            GrammarGraph newNodes = processRule(rule);
-            result.getStartNode().add(newNodes.getStartNode(), weight);
-            newNodes.getEndNode().add(result.getEndNode(), 0.0f);
-        }
-
-        return result;
-    }
-
-    /**
-     * Normalize the weights. The weights should always be zero or greater. We
-     * need to convert the weights to a log probability.
-     * 
-     * @param weights
-     *            the weights to normalize
-     */
-    private List<Float> getNormalizedWeights(List<Float> weights) {
-        
-        if (weights == null) {
-            return null;
-        }
-
-        double sum = 0.0;
-        for (float weight : weights) {
-            if (weight < 0) {
-                throw new IllegalArgumentException("Negative weight " + weight);
-            }
-            sum += weight;
-        }
-
-        List<Float> normalized = new LinkedList<Float>(weights);
-
-        for (int i = 0; i < weights.size(); i++) {
-            if (sum == 0.0f) {
-                normalized.set(i, LogMath.LOG_ZERO);
-            } else {
-                normalized.set(i, logMath.linearToLog(weights.get(i) / sum));
-            }
-        }
-        return normalized;
-    }
-
-    /**
-     * Parses the given RuleSequence into a network of GrammarNodes.
-     * 
-     * @param ruleSequence
-     *            the RuleSequence to parse
-     * @return the first and last GrammarNodes of the network
-     */
-    private GrammarGraph processRuleSequence(JSGFRuleSequence ruleSequence)
-            throws JSGFGrammarException {
-
-        GrammarNode startNode = null;
-        GrammarNode endNode = null;
-        logger.fine("parseRuleSequence: " + ruleSequence);
-
-        List<JSGFRule> rules = ruleSequence.getRules();
-
-        GrammarNode lastGrammarNode = null;
-
-        // expand and connect each rule in the sequence serially
-        for (int i = 0; i < rules.size(); i++) {
-            JSGFRule rule = rules.get(i);
-            GrammarGraph newNodes = processRule(rule);
-
-            // first node
-            if (i == 0) {
-                startNode = newNodes.getStartNode();
-            }
-
-            // last node
-            if (i == (rules.size() - 1)) {
-                endNode = newNodes.getEndNode();
-            }
-
-            if (i > 0) {
-                lastGrammarNode.add(newNodes.getStartNode(), 0.0f);
-            }
-            lastGrammarNode = newNodes.getEndNode();
-        }
-
-        return new GrammarGraph(startNode, endNode);
-    }
-
-    /**
-     * Parses the given RuleTag into a network GrammarNodes.
-     * 
-     * @param ruleTag
-     *            the RuleTag to parse
-     * @return the first and last GrammarNodes of the network
-     */
-    private GrammarGraph processRuleTag(JSGFRuleTag ruleTag)
-            throws JSGFGrammarException {
-        logger.fine("parseRuleTag: " + ruleTag);
-        JSGFRule rule = ruleTag.getRule();
-        return processRule(rule);
-    }
-
-    /**
-     * Creates a GrammarNode with the word in the given RuleToken.
-     * 
-     * @param ruleToken
-     *            the RuleToken that contains the word
-     * @return a GrammarNode with the word in the given RuleToken
-     */
-    private GrammarGraph processRuleToken(JSGFRuleToken ruleToken) {
-
-        GrammarNode node = createGrammarNode(ruleToken.getText());
-        return new GrammarGraph(node, node);
     }
 
     // ///////////////////////////////////////////////////////////////////
@@ -610,6 +227,7 @@ public class JSGFGrammar extends Grammar {
      * @throws JSGFGrammarException other exception occurred
      * @throws IOException exception during IO
      */
+    @Override
     public void commitChanges() throws IOException, JSGFGrammarParseException,
             JSGFGrammarException {
         try {
@@ -660,7 +278,7 @@ public class JSGFGrammar extends Grammar {
     /**
      * Load grammars imported by the specified RuleGrammar if they are not
      * already loaded.
-     * 
+     *
      * @throws JSGFGrammarParseException
      */
     private void loadImports(JSGFRuleGrammar grammar) throws IOException,
@@ -690,7 +308,7 @@ public class JSGFGrammar extends Grammar {
 
     /**
      * Load named grammar from import rule
-     * 
+     *
      * @param grammarName
      * @return already loaded grammar
      * @throws JSGFGrammarParseException
@@ -710,10 +328,9 @@ public class JSGFGrammar extends Grammar {
     /**
      * Load grammars imported by a fully qualified Rule Token if they are not
      * already loaded.
-     * 
+     *
      * @param grammar
      * @throws IOException
-     * @throws GrammarException
      * @throws JSGFGrammarParseException
      */
     private void loadFullQualifiedRules(JSGFRuleGrammar grammar)
@@ -750,125 +367,5 @@ public class JSGFGrammar extends Grammar {
         }
     }
 
-    /**
-     * Gets the fully resolved rule name
-     * 
-     * @param ruleName
-     *            the partial name
-     * @return the fully resolved name
-     * @throws JSGFGrammarException
-     */
-    private String getFullRuleName(String ruleName) throws JSGFGrammarException {
-        JSGFRuleName rname = ruleGrammar.resolve(new JSGFRuleName(ruleName));
-        return rname.getRuleName();
-    }
 
-    /** Dumps interesting things about this grammar */
-    protected void dumpGrammar() {
-        System.out.println("Imported rules { ");
-
-        for (JSGFRuleName imp : ruleGrammar.getImports()) {
-            System.out.println("  Import " + imp.getRuleName());
-        }
-        System.out.println("}");
-
-        System.out.println("Rulenames { ");
-
-        for (String name : ruleGrammar.getRuleNames()) {
-            System.out.println("  Name " + name);
-        }
-        System.out.println("}");
-    }
-
-    /**
-     * Represents a graph of grammar nodes. A grammar graph has a single
-     * starting node and a single ending node
-     */
-    class GrammarGraph {
-
-        private GrammarNode startNode;
-        private GrammarNode endNode;
-
-        /**
-         * Creates a grammar graph with the given nodes
-         * 
-         * @param startNode
-         *            the staring node of the graph
-         * @param endNode
-         *            the ending node of the graph
-         */
-        GrammarGraph(GrammarNode startNode, GrammarNode endNode) {
-            this.startNode = startNode;
-            this.endNode = endNode;
-        }
-
-        /** Creates a graph with non-word nodes for the start and ending nodes */
-        GrammarGraph() {
-            startNode = createGrammarNode(false);
-            endNode = createGrammarNode(false);
-        }
-
-        /**
-         * Gets the starting node
-         * 
-         * @return the starting node for the graph
-         */
-        GrammarNode getStartNode() {
-            return startNode;
-        }
-
-        /**
-         * Gets the ending node
-         * 
-         * @return the ending node for the graph
-         */
-        GrammarNode getEndNode() {
-            return endNode;
-        }
-    }
-
-    /** Manages a stack of grammar graphs that can be accessed by grammar name */
-    class RuleStack {
-
-        private List<String> stack;
-        private HashMap<String, GrammarGraph> map;
-
-        /** Creates a name stack */
-        public RuleStack() {
-            clear();
-        }
-
-        /** Pushes the grammar graph on the stack */
-        public void push(String name, GrammarGraph g) {
-            stack.add(0, name);
-            map.put(name, g);
-        }
-
-        /** remove the top graph on the stack */
-        public void pop() {
-            map.remove(stack.remove(0));
-        }
-
-        /**
-         * Checks to see if the stack contains a graph with the given name
-         * 
-         * @param name
-         *            the graph name
-         * @return the grammar graph associated with the name if found,
-         *         otherwise null
-         */
-        public GrammarGraph contains(String name) {
-            if (stack.contains(name)) {
-                return map.get(name);
-            } else {
-                return null;
-            }
-        }
-
-        /** Clears this name stack */
-        public void clear() {
-            stack = new LinkedList<String>();
-            map = new HashMap<String, GrammarGraph>();
-        }
-    }
 }
